Expose single receiver/windowseventlog config on all OSes
