+++ receiver/windowseventlogreceiver/receiver.go
@@ -1,0 +1,30 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package windowseventlogreceiver // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/windowseventlogreceiver"
+
+import (
+	"go.opentelemetry.io/collector/component"
+
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/consumerretry"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/adapter"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/input/windows"
+)
+
+// createDefaultConfig creates a config with type and version
+func createDefaultConfig() component.Config {
+	return &WindowsLogConfig{
+		BaseConfig: adapter.BaseConfig{
+			Operators:      []operator.Config{},
+			RetryOnFailure: consumerretry.NewDefaultConfig(),
+		},
+		InputConfig: *windows.NewConfig(),
+	}
+}
+
+// WindowsLogConfig defines configuration for the windowseventlog receiver
+type WindowsLogConfig struct {
+	InputConfig        windows.Config `mapstructure:",squash"`
+	adapter.BaseConfig `mapstructure:",squash"`
+}
