+++ receiver/windowseventlogreceiver/receiver_others.go
@@ -14,8 +14,6 @@ import (
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/receiver"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/adapter"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/receiver/windowseventlogreceiver/internal/metadata"
 )
 
@@ -27,14 +25,6 @@ func newFactoryAdapter() receiver.Factory {
 		receiver.WithLogs(createLogsReceiver, metadata.LogsStability))
 }
 
-func createDefaultConfig() component.Config {
-	return &WindowsLogConfig{
-		BaseConfig: adapter.BaseConfig{
-			Operators: []operator.Config{},
-		},
-	}
-}
-
 func createLogsReceiver(
 	_ context.Context,
 	_ receiver.CreateSettings,
@@ -42,9 +32,4 @@ func createLogsReceiver(
 	_ consumer.Logs,
 ) (receiver.Logs, error) {
 	return nil, fmt.Errorf("windows eventlog receiver is only supported on Windows")
-}
-
-// WindowsLogConfig defines configuration for the windowseventlog receiver
-type WindowsLogConfig struct {
-	adapter.BaseConfig `mapstructure:",squash"`
 }
